Add files via upload
The Online Shopping Store System is a C++ program that simulates an e-commerce platform. It allows users to browse, sort, and search for products such as phones, laptops, and smartwatches. Additionally, the system provides functionalities for adding items to a cart and completing orders through a payment system. The project showcases the use of structures, arrays, sorting, searching algorithms, and user interaction in C++.